requirements_dev.txt
ipython
ipdb
pylint
pylint-django
rope
black